// src/shared/api/model/personalitiesControllerFindAllParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.2.2
 */
import type { PersonalitiesControllerFindAllOrder } from "./personalitiesControllerFindAllOrder";
import type { PersonalitiesControllerFindAllSort } from "./personalitiesControllerFindAllSort";

export type PersonalitiesControllerFindAllParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * 정렬할 필드명
   */
  sort?: PersonalitiesControllerFindAllSort;
  order?: PersonalitiesControllerFindAllOrder;
  /**
   * name 검색어 (부분 일치)
   */
  search?: string;
  /**
   * jobTypeId 필터
   * @minimum 1
   */
  jobTypeId?: number;
};