// src/shared/api/model/createPersonalityDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.2.0
 */

export interface CreatePersonalityDto {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  description: string;
  /** @minimum 1 */
  jobTypeId?: number;
  /**
   * @minLength 1
   * @maxLength 500
   */
  imageUrl?: string;
}